Let ringRadiusPx default
