+++ src/display/MovablePointDisplay.test.ts
@@ -1,0 +1,7 @@
+import { MovablePointDisplay } from "./MovablePointDisplay"
+
+describe("MovablePointDisplay", () => {
+  it("has a human-readable displayName", () => {
+    expect(MovablePointDisplay.displayName).toBe("MovablePointDisplay")
+  })
+})
