+++ src/interaction/MovablePoint.test.ts
@@ -1,13 +1,7 @@
-import { MovablePoint, MovablePointSVG } from "./MovablePoint"
+import { MovablePoint } from "./MovablePoint"
 
 describe("MovablePoint", () => {
-    it("has a human-readable displayName", () => {
-        expect(MovablePoint.displayName).toBe("MovablePoint");
-    })
-})
-
-describe("MovablePointSVG", () => {
-    it("has a human-readable displayName", () => {
-        expect(MovablePointSVG.displayName).toBe("MovablePointSVG");
-    })
+  it("has a human-readable displayName", () => {
+    expect(MovablePoint.displayName).toBe("MovablePoint")
+  })
 })
